Rename alter_function_* to alter_scalar_function_*
