+++ src/catalog/catalog_entry/scalar_function_catalog_entry.cpp
@@ -1,5 +1,5 @@
 #include "duckdb/catalog/catalog_entry/scalar_function_catalog_entry.hpp"
-#include "duckdb/parser/parsed_data/alter_function_info.hpp"
+#include "duckdb/parser/parsed_data/alter_scalar_function_info.hpp"
 
 namespace duckdb {
 
@@ -9,14 +9,15 @@ ScalarFunctionCatalogEntry::ScalarFunctionCatalogEntry(Catalog *catalog, SchemaC
 }
 
 unique_ptr<CatalogEntry> ScalarFunctionCatalogEntry::AlterEntry(ClientContext &context, AlterInfo *info) {
-	if (info->type != AlterType::ALTER_FUNCTION) {
+	if (info->type != AlterType::ALTER_SCALAR_FUNCTION) {
 		throw InternalException("Attempting to alter ScalarFunctionCatalogEntry with unsupported alter type");
 	}
-	auto &function_info = (AlterFunctionInfo &)*info;
-	if (function_info.alter_function_type != AlterFunctionType::ADD_FUNCTION_OVERLOADS) {
-		throw InternalException("Attempting to alter ScalarFunctionCatalogEntry with unsupported alter function type");
+	auto &function_info = (AlterScalarFunctionInfo &)*info;
+	if (function_info.alter_scalar_function_type != AlterScalarFunctionType::ADD_FUNCTION_OVERLOADS) {
+		throw InternalException(
+		    "Attempting to alter ScalarFunctionCatalogEntry with unsupported alter scalar function type");
 	}
-	auto &add_overloads = (AddFunctionOverloadInfo &)function_info;
+	auto &add_overloads = (AddScalarFunctionOverloadInfo &)function_info;
 
 	ScalarFunctionSet new_set = functions;
 	if (!new_set.MergeFunctionSet(add_overloads.new_overloads)) {
