+++ src/include/duckdb/parser/parsed_data/alter_info.hpp
@@ -20,7 +20,8 @@ enum class AlterType : uint8_t {
 	ALTER_VIEW = 2,
 	ALTER_SEQUENCE = 3,
 	CHANGE_OWNERSHIP = 4,
-	ALTER_FUNCTION = 5
+	ALTER_SCALAR_FUNCTION = 5,
+	ALTER_TABLE_FUNCTION = 6
 };
 
 struct AlterEntryData {
